Revert "kmod: Upload static configuration to switch before connecting the PHYs."
Actually we are still uploading the static config before phy_connect.
But this change has taken the config_upload function out of parse_dt,
(along with the phy_connect, in fact) just as it was prior to the
reverted commit. The situation is a bit more complicated now that we
need to perform hot-patching to the static config, based on the parsed
DTS, but also before said static config is uploaded to the hardware.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>